// generated/src/app/api-models/conversion-factor-dtl.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class ConversionFactorDtl {
  seqConvFactor: number;
  convFactorId: string;
  factor: number;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}